Split the view and environment into their own files
