+++ src/h_assets/__init__.py
@@ -1,1 +1,2 @@
-from h_assets.assets import Environment, assets_view
+from h_assets.environment import Environment
+from h_assets.view import assets_view
