+++ tests/unit/h_assets/environment_test.py
@@ -1,11 +1,9 @@
 from configparser import ConfigParser
-from unittest.mock import create_autospec, sentinel
+from unittest.mock import sentinel
 
 import pytest
-from pyramid.httpexceptions import HTTPNotFound
-from pyramid.testing import DummyRequest
 
-from h_assets.assets import Environment, assets_view
+from h_assets.environment import Environment
 
 
 class TestEnvironment:
@@ -75,7 +73,7 @@ class TestEnvironment:
 
     @pytest.fixture(autouse=True)
     def CachedINIFile(self, patch):
-        CachedINIFile = patch("h_assets.assets.CachedINIFile")
+        CachedINIFile = patch("h_assets.environment.CachedINIFile")
 
         parser = ConfigParser()
         parser.read_dict(
@@ -92,7 +90,7 @@ class TestEnvironment:
 
     @pytest.fixture(autouse=True)
     def CachedJSONFile(self, patch):
-        CachedJSONFile = patch("h_assets.assets.CachedJSONFile")
+        CachedJSONFile = patch("h_assets.environment.CachedJSONFile")
 
         CachedJSONFile.return_value.load.return_value = {
             "app.bundle.js": "app.bundle.js?hash_app",
@@ -100,38 +98,3 @@ class TestEnvironment:
         }
 
         return CachedJSONFile
-
-
-class TestAssetsView:
-    def test_it_returns_static_view_response_if_cache_buster_valid(
-        self, static_view, environment
-    ):
-        environment.check_cache_buster.return_value = True
-        request = DummyRequest(query_string=sentinel.query)
-        request.path = "/path"
-
-        response = assets_view(environment)(sentinel.context, request)
-
-        environment.check_cache_buster.assert_called_once_with("/path", sentinel.query)
-
-        static_view.assert_called_once_with(
-            environment.asset_root.return_value, cache_max_age=None, use_subpath=True
-        )
-        static_view.return_value.assert_called_with(sentinel.context, request)
-        assert response == static_view.return_value.return_value
-
-    def test_it_returns_404_if_cache_buster_invalid(self, environment, static_view):
-        environment.check_cache_buster.return_value = False
-
-        response = assets_view(environment)({}, DummyRequest())
-
-        static_view.return_value.assert_not_called()
-        assert isinstance(response, HTTPNotFound)
-
-    @pytest.fixture
-    def environment(self):
-        return create_autospec(Environment, instance=True, spec_set=True)
-
-    @pytest.fixture(autouse=True)
-    def static_view(self, patch):
-        return patch("h_assets.assets.static_view")
